Add files via upload
Fixed issue with definition of stats window at the end of files. This improper definition could lead to a crash in previous versions, depending on the length of the file and the window settings. 
Set cleanYwith parameter default to 'no' in pop_iCanClean to match GUI default This prevents minor issue related to eegh command and setting the reference signals to be a copy [filtered] version of the EEG data itself.
Added check that at least 2 rows of data are provided for X and Y before the canoncorr function is called.
Added some comments.